requirements.txt
pandas
requests
jupyter
notebook
psycopg2-binary
pyspark
